Projekt/requirements.txt
Flask==2.3.2
paho-mqtt==1.6.1
openaq==1.1.0
